rand_xorshift

Struct XorShiftRng

pub struct XorShiftRng { /* private fields */ }

An Xorshift random number generator.

The Xorshift1 algorithm is not suitable for cryptographic purposes but is very fast. If you do not know for sure that it fits your requirements, use a more secure one such as StdRng or OsRng.

When seeded with zero (i.e. XorShiftRng::from_seed(0) is called), this implementation actually uses 0xBAD_5EED_0BAD_5EED_0BAD_5EED_0BAD_5EED for the seed. This arbitrary value is used because the underlying algorithm canโt escape from an all-zero state, and the function is infallible so it canโt signal this by returning an error.

Marsaglia, George (July 2003). โXorshift RNGsโ. Journal of Statistical Software. Vol. 8 (Issue 14). โฉ